test more things in build
